Alteradas classes Repository e Service de Ocorrencia e criada classe Controller da mesma entidade

+++ Aplicação/CID/src/main/java/br/edu/unoesc/CID/entity/Ocorrencia.java
@@ -14,7 +14,7 @@ public class Ocorrencia {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected int idOcorrencia;
+    protected long idOcorrencia;
 
     @Column(name = "descoco")
     protected String descricaoOcorrencia;
